run Release Charts workflow only if Chart.yaml has been modified
Previously the workflow simply was triggered on pushed to the `main` branch. Which is fine, but will naturally fail, if the `Chart.yaml` has not been modified and there the to-be-released version already has been released.

This change will primarily watch for changes made to `Chart.yaml`, then verify if it is indeed the `main` branch and if both conditions are met, it will go forward with creating a new chart release.